+++ .vscode/settings.json
@@ -1,3 +1,8 @@
-{
+[
+  {
     "java.configuration.updateBuildConfiguration": "automatic"
-}
+  },
+  {
+    "java.refactor.renameFromFileExplorer": "autoApply"
+  }
+]
