+++ University_Simulator/src/main/java/com/challenge/alkemy/repositories/SubjectRepository.java
@@ -1,4 +1,6 @@
-package com.challenge.alkemy;
+package com.challenge.alkemy.repositories;
+
+import com.challenge.alkemy.Subject;
 
 import org.springframework.data.repository.CrudRepository;
